Move plugins to common place since they are common for now
