+++ kmp/kmp-module-b/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("kmp-platform-plugin")
+    id("javascript-platform-plugin")
 }
 
 kotlin {
